Inserting the response into Db and then showing the result from db into different fragment.

+++ app/src/main/java/com/nikhil/weatherapp/database/WeatherDetailsDao.kt
@@ -9,7 +9,7 @@ import io.reactivex.Single
 interface WeatherDetailsDao {
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
-    fun insert(weatherDetails: WeatherEntity): Single<Int>
+    fun insert(weatherDetails: WeatherEntity): Single<Long>
 
     @Query("select * from weather_table")
     fun getAllData(): LiveData<List<WeatherEntity>>
